Function extracted so we can have browser-run code in its own file
